Move `LoginUserWithOtp` to the login package
