+++ app/src/androidTest/java/org/simple/clinic/di/TestAppComponent.kt
@@ -29,7 +29,7 @@ import org.simple.clinic.security.pin.BruteForceProtectionAndroidTest
 import org.simple.clinic.storage.DaoWithUpsertAndroidTest
 import org.simple.clinic.storage.files.AndroidFileStorageAndroidTest
 import org.simple.clinic.summary.addphone.MissingPhoneReminderRepositoryAndroidTest
-import org.simple.clinic.user.LoginUserWithOtpServerIntegrationTest
+import org.simple.clinic.login.LoginUserWithOtpServerIntegrationTest
 import org.simple.clinic.user.OngoingLoginEntryRepositoryTest
 import org.simple.clinic.user.UserDaoAndroidTest
 import org.simple.clinic.user.UserSessionAndroidTest
