+++ app/src/test/java/org/simple/clinic/login/LoginUserWithOtpTest.kt
@@ -1,4 +1,4 @@
-package org.simple.clinic.user
+package org.simple.clinic.login
 
 import com.f2prateek.rx.preferences2.Preference
 import com.google.common.truth.Truth.assertThat
@@ -20,16 +20,14 @@ import org.junit.Test
 import org.simple.clinic.analytics.Analytics
 import org.simple.clinic.analytics.MockAnalyticsReporter
 import org.simple.clinic.facility.FacilityRepository
-import org.simple.clinic.login.LoginApi
-import org.simple.clinic.login.LoginRequest
-import org.simple.clinic.login.LoginResponse
 import org.simple.clinic.login.LoginResult.NetworkError
 import org.simple.clinic.login.LoginResult.ServerError
 import org.simple.clinic.login.LoginResult.Success
 import org.simple.clinic.login.LoginResult.UnexpectedError
-import org.simple.clinic.login.UserPayload
 import org.simple.clinic.patient.PatientMocker
 import org.simple.clinic.sync.DataSync
+import org.simple.clinic.user.LoggedInUserPayload
+import org.simple.clinic.user.User
 import org.simple.clinic.util.Just
 import org.simple.clinic.util.Optional
 import org.simple.clinic.util.RxErrorsRule
